refactor(test): split execution service tests by theme
